Enabling System.Runtime.Caching on Linux: - split platform dependent classes to partial ones and updated the project - updated tests to be able to run them on Unix - addressing feedback - fixing build issues on Unix
