+++ src/System.Runtime.Caching/src/System/Runtime/Caching/MemoryMonitor.cs
@@ -16,7 +16,7 @@ namespace System.Runtime.Caching
     // drop cache entries to avoid paging.  The second monitors the amount of memory used by
     // the cache itself, and helps determine when we should drop cache entries to avoid 
     // exceeding the cache's memory limit.  Both are configurable (see ConfigUtil.cs).
-    internal abstract class MemoryMonitor
+    internal abstract partial class MemoryMonitor
     {
         protected const int TERABYTE_SHIFT = 40;
         protected const long TERABYTE = 1L << TERABYTE_SHIFT;
@@ -39,25 +39,11 @@ namespace System.Runtime.Caching
         protected int[] _pressureHist;
         protected int _pressureTotal;
 
-        private static long s_totalPhysical;
-        private static long s_totalVirtual;
+        private static long s_totalPhysical = 0;
+        private static long s_totalVirtual = 0;
 
-        static MemoryMonitor()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                Interop.Kernel32.MEMORYSTATUSEX memoryStatusEx = default;
-                memoryStatusEx.dwLength = (uint)Marshal.SizeOf<Interop.Kernel32.MEMORYSTATUSEX>();
-                if (Interop.Kernel32.GlobalMemoryStatusEx(out memoryStatusEx) != 0)
-                {
-                    s_totalPhysical = (long)memoryStatusEx.ullTotalPhys;
-                    s_totalVirtual = (long)memoryStatusEx.ullTotalVirtual;
-                }
-            }
-        }
-
-        internal static long TotalPhysical { get { return s_totalPhysical; } }
-        internal static long TotalVirtual { get { return s_totalVirtual; } }
+        internal static long TotalPhysical => s_totalPhysical;
+        internal static long TotalVirtual => s_totalVirtual;
 
         internal int PressureLast { get { return _pressureHist[_i0]; } }
         internal int PressureHigh { get { return _pressureHigh; } }
